Refactor code to support querying certain fields instead of all only
Code has been refactored to support querying certain fields from a class. Before when you tried to let TinyMapper map a resultset to an object. It would try to set all fields that were annotated in the object, whether these were being queried or not. This meant it only worked if you used 'SELECT * FROM x' aka if you queried all fields.
The only downside atm is that the infinite embedding doesn't work anymore. At the moment it is only possible to embed 2 levels down.